heater control vlues to firebase
Teperature activity, which contains heater control switch was modified so that the click of a switch sends a new value to firebase database heatSwitchEvents(id, dataTime, state) which records each time the switch was triggered on and off.
Next: to read last value from the database to set the condition of the switch;
next: create an extra activity to display the log of heater switch events and corresponding dates and times.